requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
firebase-admin==6.5.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.7
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.0.8
requests==2.32.3
rsa==4.9
setuptools==74.0.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.10.0
